new RevisionInfo class obsoletes anonymous objRevInfo class
obsolete getObjRevInfo() in Ui\Recent and Ui\Revisions classes, that had defined anonymous class. they are similar but slightly differently treat the diff link icon. It is impossible to create common method for diff icon using only revision information, it depends on caller, ie. recent or revisions list
