+++ inc/Ui/MediaRevisions.php
@@ -3,6 +3,7 @@
 namespace dokuwiki\Ui;
 
 use dokuwiki\ChangeLog\MediaChangeLog;
+use dokuwiki\ChangeLog\RevisionInfo;
 use dokuwiki\Form\Form;
 use InvalidArgumentException;
 
@@ -84,16 +85,16 @@ class MediaRevisions extends Revisions
             }
             $form->addHTML(' ');
 
-            $objRevInfo = $this->getObjRevInfo($info);
+            $RevInfo = new RevisionInfo($info);
             $html = implode(' ', [
-                $objRevInfo->editDate(),          // edit date and time
-                $objRevInfo->difflink(),          // link to diffview icon
-                $objRevInfo->itemName(),          // name of page or media
+                $RevInfo->editDate(),          // edit date and time
+                $RevInfo->difflink(),          // link to diffview icon
+                $RevInfo->itemName(),          // name of page or media
                 '<div>',
-                $objRevInfo->editSummary(),       // edit summary
-                $objRevInfo->editor(),            // editor info
-                $objRevInfo->sizechange(),        // size change indicator
-                $objRevInfo->currentIndicator(),  // current indicator (only when k=1)
+                $RevInfo->editSummary(),       // edit summary
+                $RevInfo->editor(),            // editor info
+                $RevInfo->sizechange(),        // size change indicator
+                $RevInfo->currentIndicator(),  // current indicator (only when k=1)
                 '</div>',
             ]);
             $form->addHTML($html);
